+++ CHANGELOG.md
@@ -4,7 +4,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/orca-services/cakephp-heartbeat)
 ### Added
-- Add JSON view 
 
 ### Changed
 
@@ -15,6 +14,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Dependencies
 
+### Fixed
+
+## [4.0.3](https://github.com/orca-services/cakephp-heartbeat/releases/tag/4.0.2) - 2026-09-17
 ### Fixed
 - Re-enable JSON view for heartbeat status page